requirements.txt
# Gauge Test Environment - Minimal Dependencies
nicegui>=2.21.0
Pillow>=10.0.0

